requirements.txt
djangorestframework
psycop2-binary